Membagi Question page jadi beberapa komponent
